#284 refactor: Create usePortfolioAddOrEditDialogInputs
